Move all gear components into root gear folder.
