+++ src/components/gear/gear-equipment-form/gear-equipment-form.html
