+++ src/components/gear/gear-list-filter-bar/gear-list-filter-bar.html
