+++ src/components/gear/gear-list/gear-list.html
