+++ src/components/gear/gear-list/gear-list.ts
@@ -1,28 +1,28 @@
 import html from "./gear-list.html";
-import { GetAllGearFeature } from "../../features/gear/get-all-gear/get-all-gear-feature";
-import { GearListItem } from "../../features/gear/gear-list-item";
-import { EmptyRequest } from "../../lib/common/features/empty-request";
-import { EventBus } from "../../lib/events/event-bus";
-import { GearFilterChangedEvent } from "../../lib/events/gear-filter-changed-event";
-import { AppEvent } from "../../lib/events/app-event";
-import { FilterGearListFeature } from "../../features/gear/filter-gear-list/filter-gear-list-feature";
-import { FilterGearListRequest } from "../../features/gear/filter-gear-list/filter-gear-list-request";
-import { EventType } from "../../lib/events/event-type";
-import { SortGearListFeature } from "../../features/gear/sort-gear-list/sort-gear-list-feature";
-import { SortGearListRequest } from "../../features/gear/sort-gear-list/sort-gear-list-request";
+import { GetAllGearFeature } from "../../../features/gear/get-all-gear/get-all-gear-feature";
+import { GearListItem } from "../../../features/gear/gear-list-item";
+import { EmptyRequest } from "../../../lib/common/features/empty-request";
+import { EventBus } from "../../../lib/events/event-bus";
+import { GearFilterChangedEvent } from "../../../lib/events/gear-filter-changed-event";
+import { AppEvent } from "../../../lib/events/app-event";
+import { FilterGearListFeature } from "../../../features/gear/filter-gear-list/filter-gear-list-feature";
+import { FilterGearListRequest } from "../../../features/gear/filter-gear-list/filter-gear-list-request";
+import { EventType } from "../../../lib/events/event-type";
+import { SortGearListFeature } from "../../../features/gear/sort-gear-list/sort-gear-list-feature";
+import { SortGearListRequest } from "../../../features/gear/sort-gear-list/sort-gear-list-request";
 import { GearEquipmentFormComponent } from "../gear-equipment-form/gear-equipment-form";
-import { ModalDialogComponent } from "../modal-dialog/modal-dialog";
-import { EquipmentItem } from "../../features/gear/equipment-item";
-import { GetGearByIdFeature } from "../../features/gear/get-gear-by-id/get-gear-by-id-feature";
-import { GetGearByIdRequest } from "../../features/gear/get-gear-by-id/get-gear-by-id-request";
-import { GearItem } from "../../features/gear/gear-item";
+import { ModalDialogComponent } from "../../modal-dialog/modal-dialog";
+import { EquipmentItem } from "../../../features/gear/equipment-item";
+import { GetGearByIdFeature } from "../../../features/gear/get-gear-by-id/get-gear-by-id-feature";
+import { GetGearByIdRequest } from "../../../features/gear/get-gear-by-id/get-gear-by-id-request";
+import { GearItem } from "../../../features/gear/gear-item";
 import { GearArmorFormComponent } from "../gear-armor-form/gear-armor-form";
-import { ArmorItem } from "../../features/gear/armor-item";
-import { WeaponItem } from "../../features/gear/weapon-item";
+import { ArmorItem } from "../../../features/gear/armor-item";
+import { WeaponItem } from "../../../features/gear/weapon-item";
 import { GearWeaponFormComponent } from "../gear-weapon-form/gear-weapon-form";
-import { AppEventListener } from "../../lib/events/app-event-listener-interface";
-import { BaseListComponent } from "../base-list/base-list-component";
-import { TableHeader } from "../../lib/tables/table-header";
+import { AppEventListener } from "../../../lib/events/app-event-listener-interface";
+import { BaseListComponent } from "../../base-list/base-list-component";
+import { TableHeader } from "../../../lib/tables/table-header";
 
 export class GearListComponent extends BaseListComponent {
     private gearList: Array<GearListItem> = [];
