+++ src/components/gear/gear-weapon-form/gear-weapon-form.html
